se arreglaron imagenes

+++ app/Http/Controllers/ProfileController.php
@@ -40,6 +40,13 @@ class ProfileController extends Controller
         return view('profile.create-consulta-cliente');
     }
 
+    /**
+     *
+     *
+     *
+     * @param  \Illuminate\Http\Request
+     * @return \Illuminate\Http\Response
+     */
     public function storeConsultaCliente(Request $request)
     {
         $authenticatedUser = Auth::user();
@@ -83,12 +90,14 @@ class ProfileController extends Controller
         $user->fill($request->validated());
 
 
+        $user->fill($request->validated());
+
         if ($request->hasFile('imagen')) {
-            if ($user->imagen) {
+            if ($user->imagen && $user->imagen != 'icon-person.jpg') {
                 Storage::disk('public')->delete('images/' . $user->imagen);
             }
-            $path = $request->file('imagen')->store('images', 'public');
-            $user->imagen = basename($path);
+            $imagenPath = $request->file('imagen')->store('images', 'public');
+            $user->imagen = basename($imagenPath);  // Aquí sí
         }
